apps/api-gateway/requirements.txt
fastapi==0.128.0
gunicorn==23.0.0
uvicorn==0.18.2
prometheus_client==0.24.1
python_json_logger==4.0.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-instrumentation-fastapi==0.60b1